Fix moving from my library
